+++ packages/app/src/app/components/modals/manage-tags/manage-tags.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, inject, signal } from '@angular/core';
 import { FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { TranslatePipe } from '@ngx-translate/core';
+import { ConfirmService } from '../../../services/confirm.service';
 import { QbService } from '../../../services/qb.service';
 import { ServerStoreService } from '../../../services/server-store.service';
 
@@ -15,6 +16,7 @@ import { ServerStoreService } from '../../../services/server-store.service';
 export class ManageTags implements OnInit {
   private readonly qbService = inject(QbService);
   private readonly serverStoreService = inject(ServerStoreService);
+  private readonly confirmService = inject(ConfirmService);
   public readonly activeModal = inject(NgbActiveModal);
 
   public tags = signal<string[]>([]);
@@ -52,6 +54,16 @@ export class ManageTags implements OnInit {
   }
 
   public async delete(tag: string): Promise<void> {
+    const confirmed = await this.confirmService.confirm(
+      'components.modals.manage-tags.delete-confirm.title',
+      {
+        text: 'components.modals.manage-tags.delete-confirm.message',
+        data: { name: tag },
+      },
+      'general.button.delete',
+    );
+    if (!confirmed) return;
+
     const serverId = this.serverStoreService.currentServerId() as string;
     try {
       await this.qbService.deleteTags(serverId, [tag]);
